add TargetSelectionPage, remove AccountSelectionPage, fix saveAndClose bugs
